+++ example/widgets/line_chart_card.dart
@@ -65,4 +65,3 @@ class LineChartCard extends StatelessWidget {
     );
   }
 }
-